copy all modules if available
For development it can be helpful to include all kernel modules in the
initramfs and copy them to the rootfs during boot. Add support for this
directly in the initramfs.

This relies on the fact that development kernels include the git tag and
build number (incrementing) so 'uname -r' changes every time.

It's up to the user to manage deleting their old modules.

Opt-in to this behaviour by setting
deviceinfo_dev_kernel_copy_all_modules="true" in your deviceinfo file.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>