main/postmarketos-initramfs: add option to force loading modules
This brings back load_modules from the days when we used mdev, and calls
it before udev. If for some reason udev can't load modules, this at
least gives some sort of workaround for devices that can't boot with
udev-only [1]. This helps with debugging issues brought about by
7c981277, and gives device owners a way to temporarily work around it
until the kernel can be fixed.

The feature is hidden behind a kernel cmdline arg,
pmos.force-load-modules

1. https://gitlab.com/postmarketOS/pmaports/-/merge_requests/5287