enable EFI on sdm845 phones
Enable support for generating systemd-boot configs on SDM845 devices
while continuing to support the Android boot image.

This additionally switches over the flash method to fastboot-bootpart,
so the pmOS boot and root partitions are exported seperately. The
expectation is that users STOP using the subpartitioning scheme and
flash the two partitions separately.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>